Move wireguard cfg to a leaf package without deps
So that supervisor can using without pulling all the bad stuff
