+++ mobile/mysterium/wireguard_connection_setup_test.go
@@ -28,6 +28,7 @@ import (
 	"github.com/mysteriumnetwork/node/core/connection"
 	"github.com/mysteriumnetwork/node/core/ip"
 	wg "github.com/mysteriumnetwork/node/services/wireguard"
+	"github.com/mysteriumnetwork/node/services/wireguard/wgcfg"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -135,14 +136,14 @@ func (m mockWireGuardDevice) Start(_ string, _ wg.ServiceConfig, _ *net.UDPConn)
 func (m mockWireGuardDevice) Stop() {
 }
 
-func (m mockWireGuardDevice) Stats() (*wg.Stats, error) {
-	return &wg.Stats{BytesSent: 10, BytesReceived: 11}, nil
+func (m mockWireGuardDevice) Stats() (*wgcfg.Stats, error) {
+	return &wgcfg.Stats{BytesSent: 10, BytesReceived: 11}, nil
 }
 
 type mockHandshakeWaiter struct {
 	err error
 }
 
-func (m *mockHandshakeWaiter) Wait(statsFetch func() (*wg.Stats, error), timeout time.Duration, stop <-chan struct{}) error {
+func (m *mockHandshakeWaiter) Wait(statsFetch func() (*wgcfg.Stats, error), timeout time.Duration, stop <-chan struct{}) error {
 	return m.err
 }
