+++ services/wireguard/connection/connection_test.go
@@ -28,6 +28,7 @@ import (
 	"github.com/mysteriumnetwork/node/core/connection"
 	"github.com/mysteriumnetwork/node/core/ip"
 	wg "github.com/mysteriumnetwork/node/services/wireguard"
+	"github.com/mysteriumnetwork/node/services/wireguard/wgcfg"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -135,18 +136,18 @@ func (mce *mockConnectionEndpoint) StartProviderMode(config wg.ProviderModeConfi
 func (mce *mockConnectionEndpoint) InterfaceName() string                                { return "mce0" }
 func (mce *mockConnectionEndpoint) Stop() error                                          { return nil }
 func (mce *mockConnectionEndpoint) Config() (wg.ServiceConfig, error)                    { return wg.ServiceConfig{}, nil }
-func (mce *mockConnectionEndpoint) AddPeer(_ string, _ wg.Peer) error                    { return nil }
+func (mce *mockConnectionEndpoint) AddPeer(_ string, _ wgcfg.Peer) error                 { return nil }
 func (mce *mockConnectionEndpoint) RemovePeer(_ string) error                            { return nil }
 func (mce *mockConnectionEndpoint) ConfigureRoutes(_ net.IP) error                       { return nil }
-func (mce *mockConnectionEndpoint) PeerStats() (*wg.Stats, error) {
-	return &wg.Stats{LastHandshake: time.Now(), BytesSent: 10, BytesReceived: 11}, nil
+func (mce *mockConnectionEndpoint) PeerStats() (*wgcfg.Stats, error) {
+	return &wgcfg.Stats{LastHandshake: time.Now(), BytesSent: 10, BytesReceived: 11}, nil
 }
 
 type mockHandshakeWaiter struct {
 	err error
 }
 
-func (m *mockHandshakeWaiter) Wait(statsFetch func() (*wg.Stats, error), timeout time.Duration, stop <-chan struct{}) error {
+func (m *mockHandshakeWaiter) Wait(statsFetch func() (*wgcfg.Stats, error), timeout time.Duration, stop <-chan struct{}) error {
 	return m.err
 }
