+++ services/wireguard/connection/handshake.go
@@ -23,13 +23,13 @@ import (
 	"net"
 	"time"
 
-	"github.com/mysteriumnetwork/node/services/wireguard"
+	"github.com/mysteriumnetwork/node/services/wireguard/wgcfg"
 )
 
 // HandshakeWaiter waits for handshake.
 type HandshakeWaiter interface {
 	// Wait waits until WireGuard does initial handshake.
-	Wait(statsFetch func() (*wireguard.Stats, error), timeout time.Duration, stop <-chan struct{}) error
+	Wait(statsFetch func() (*wgcfg.Stats, error), timeout time.Duration, stop <-chan struct{}) error
 }
 
 // NewHandshakeWaiter returns handshake waiter instance.
@@ -40,7 +40,7 @@ func NewHandshakeWaiter() HandshakeWaiter {
 type handshakeWaiter struct {
 }
 
-func (h *handshakeWaiter) Wait(statsFetch func() (*wireguard.Stats, error), timeout time.Duration, stop <-chan struct{}) error {
+func (h *handshakeWaiter) Wait(statsFetch func() (*wgcfg.Stats, error), timeout time.Duration, stop <-chan struct{}) error {
 	// We need to send any packet to initialize handshake process.
 	handshakePingConn, err := net.DialTimeout("tcp", "8.8.8.8:53", 100*time.Millisecond)
 	if err == nil {
