+++ services/wireguard/service/stats_publisher.go
@@ -21,13 +21,13 @@ import (
 	"time"
 
 	"github.com/mysteriumnetwork/node/eventbus"
-	wg "github.com/mysteriumnetwork/node/services/wireguard"
+	"github.com/mysteriumnetwork/node/services/wireguard/wgcfg"
 	"github.com/mysteriumnetwork/node/session/event"
 	"github.com/rs/zerolog/log"
 )
 
 type statsSupplier interface {
-	PeerStats() (*wg.Stats, error)
+	PeerStats() (*wgcfg.Stats, error)
 }
 
 type statsPublisher struct {
