+++ supervisor/daemon/wireguard/wginterface/interface_windows.go
@@ -23,7 +23,7 @@ import (
 	"log"
 	"strings"
 
-	wg "github.com/mysteriumnetwork/node/services/wireguard"
+	"github.com/mysteriumnetwork/node/services/wireguard/wgcfg"
 	"github.com/mysteriumnetwork/node/utils/netutil"
 	"golang.zx2c4.com/wireguard/device"
 	"golang.zx2c4.com/wireguard/ipc"
@@ -31,7 +31,7 @@ import (
 )
 
 // New creates new WgInterface instance.
-func New(cfg wg.DeviceConfig, uid string) (*WgInterface, error) {
+func New(cfg wgcfg.DeviceConfig, uid string) (*WgInterface, error) {
 	log.Println("Creating Wintun interface")
 
 	wintun, err := tun.CreateTUN(cfg.IfaceName, 0)
@@ -99,7 +99,7 @@ func (a *WgInterface) Down() {
 	a.Device.Close()
 }
 
-func configureNetwork(cfg wg.DeviceConfig) error {
+func configureNetwork(cfg wgcfg.DeviceConfig) error {
 	if err := netutil.AssignIP(cfg.IfaceName, cfg.Subnet); err != nil {
 		return fmt.Errorf("failed to assign IP address: %w", err)
 	}
